[docs] Recommend PUBLIC_RELEASE
Most people following the guide are currently not really developers and this was a fairly hidden option that's easily missed.

The disabled version will later change to be an actual custom cmake target instead of a preprocessor flag, but that'll require more code to be adjusted.

Not using the flag means that Cemu keeps some (performance-costing) debug checks in, alongside more logging and a debug logging console on Windows.